+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Changed
+- The create method from the Transaction resource is now deprecated
+
 ## [1.1.0] - 2025-10-16
 ### Added
 - InvoicePullSubscription and InvoicePullRequest resources